Move prototype system to core (world-server place is bad to domain driven)
