+++ core/src/main/java/ru/finex/core/repository/impl/GameObjectComponentPrototypeRepository.java
@@ -1,10 +1,10 @@
-package ru.finex.ws.repository;
+package ru.finex.core.repository.impl;
 
 import org.hibernate.Session;
 import org.hibernate.query.Query;
 import ru.finex.core.db.impl.TransactionalContext;
+import ru.finex.core.model.entity.GameObjectComponentPrototype;
 import ru.finex.core.repository.DefaultCrudRepository;
-import ru.finex.ws.model.entity.GameObjectComponentPrototype;
 
 import java.util.List;
 import java.util.concurrent.ExecutorService;
