+++ core/src/main/resources/evolution/core_1_GameObjectPrototype.sql
