+++ core/src/main/resources/mappings/core_GameObjectComponentPrototype.xml
