+++ core/src/main/resources/mappings/core_GameObjectPrototype.xml
