+++ worldserver/src/main/java/ru/finex/ws/inject/module/loader/WorldServerModule.java
@@ -17,9 +17,9 @@ import ru.finex.core.inject.module.NetworkModule;
 import ru.finex.core.inject.module.PersistenceModule;
 import ru.finex.core.inject.module.PlaceholderJuelModule;
 import ru.finex.core.inject.module.PoolModule;
+import ru.finex.core.prototype.ComponentPrototype;
+import ru.finex.core.prototype.ComponentPrototypeMapper;
 import ru.finex.core.utils.GenericUtils;
-import ru.finex.ws.model.prototype.ComponentPrototype;
-import ru.finex.ws.model.prototype.ComponentPrototypeMapper;
 import ru.finex.ws.service.impl.ComponentServiceImpl;
 import ru.finex.ws.tick.TickService;
 import ru.finex.ws.tick.impl.RegisterTickListener;
